Fix compatibility with Symfony 2.7 and enable translation caching
The previous implementation was overwriting all options to disable
caching to avoid reusing the same cache than the normal translator for
our translator without fallback.
This does not work anymore in Symfony 2.7 as resources are also injected
through the options and so we need to preserve them.
Options are not reset entirely anymore, but the cache folder is
overwritten instead.
This also optimizes performance as the translations are cached.